lib/settings: rename to settings_helpers
