+++ applications/settings/brightness_settings/scenes/scene_main.c
@@ -1,6 +1,6 @@
 #include "../brightness_settings.h"
 #include "../models/brightness.h"
-#include <settings/settings_gui_params.h>
+#include <settings_helpers/gui_params.h>
 
 #include <gui/modules/var_item_list.h>
