+++ applications/settings/fw_update/fw_update.c
@@ -1,7 +1,7 @@
 #include "fw_update.h"
 #include "scenes/update_scenes.h"
-#include <settings/settings_app_info.h>
-#include <settings/settings_gui_params.h>
+#include <settings_helpers/app_info.h>
+#include <settings_helpers/gui_params.h>
 
 static bool fw_update_thread_signal_callback(uint32_t signal, void* arg, void* context) {
     UNUSED(arg);
