+++ applications/settings/matter_settings/matter_settings.c
@@ -1,7 +1,7 @@
 #include "matter_settings.h"
 #include "scenes/matter_scenes.h"
-#include <settings/settings_app_info.h>
-#include <settings/settings_gui_params.h>
+#include <settings_helpers/app_info.h>
+#include <settings_helpers/gui_params.h>
 
 static bool matter_settings_thread_signal_callback(uint32_t signal, void* arg, void* context) {
     UNUSED(arg);
