+++ applications/settings/matter_settings/scenes/scene_commission_done.c
@@ -1,5 +1,5 @@
 #include "../matter_settings.h"
-#include <settings/status_view.h>
+#include <settings_helpers/status_view.h>
 
 typedef struct {
     StatusView* statuses[GuiDisplayIdMax];
