+++ lib/settings_helpers/SConscript
@@ -2,17 +2,18 @@ Import("env")
 
 env.Append(
     CPPPATH=[
-        "#/lib/settings",
+        "#/lib/settings_helphers",
     ],
     LINT_SOURCES=[
         Dir("."),
     ],
     SDK_HEADERS=[
         File("settings_app_info.h"),
+        File("settings_gui_params.h"),
     ],
 )
 
-libenv = env.Clone(FW_LIB_NAME="settings")
+libenv = env.Clone(FW_LIB_NAME="settings_helpers")
 libenv.ApplyLibFlags()
 
 sources = libenv.GlobRecursive("*.c")
