+++ lib/settings_helpers/app_info.h
