+++ lib/settings_helpers/status_view.c
